Initial version of buying functionality implemented

+++ Tech Shop/Models/Device.cs
@@ -10,7 +10,7 @@ namespace Tech_Shop.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public ushort Price { get; set; }
+        public decimal Price { get; set; }
         public string shortDescription { get; set; }
         public string description { get; set; }
        // public DeviceCategory category { get; set; }
